Add files via upload
Here we provide a Python version, with the EEG application we introduced in the paper, we plot the membership matrix and CVI vs m. 
When you use the code, you can also specify a grid search for the number of clusters / the fuzziness parameter m, and choose the one with the minimum CVI value (in our code, this is provided by S_value, computed by the compute_S function.

We reported the result using our R version of code, users are free to choose which version of code they prefer.