Add files via upload
This is a Python version of the FCPCA method; the CVI value is computed by the compute_S function.

Users can do a grid search to find the optimal selection of m and the number of clusters. 

The EEG data can be found on the website: https://figshare.com/articles/dataset/EEG_driver_drowsiness_dataset/14273687
And if you want to replicate the data application, remember to change the file path.